+++ app/src/main/java/com/gianlu/pretendyourexyzzy/api/Pyx.java
@@ -118,16 +118,19 @@ public class Pyx implements Closeable {
         trace.putAttribute("params_num", String.valueOf(params.length));
 
         try {
-            Exception lastEx = null;
+            List<Exception> exceptions = new ArrayList<>(3);
             for (int i = 0; i < 3; i++) {
                 try {
-                    return requestInternal(operation, params);
+                    PyxResponse resp = requestInternal(operation, params);
+                    resp.exceptions = exceptions;
+                    return resp;
                 } catch (SocketTimeoutException ex) {
-                    lastEx = ex;
+                    exceptions.add(ex);
                     trace.incrementMetric("tries", 1);
                     Log.d(TAG, "Socket timeout, retrying.", ex);
                 } catch (PyxException ex) {
-                    lastEx = ex;
+                    exceptions.add(ex);
+                    ex.exceptions = exceptions;
 
                     trace.incrementMetric("pyx_tries", 1);
                     trace.putAttribute("pyx_retry_error", ex.errorCode);
@@ -136,6 +139,7 @@ public class Pyx implements Closeable {
                 }
             }
 
+            Exception lastEx = exceptions.get(exceptions.size() - 1);
             if (lastEx instanceof IOException) throw (IOException) lastEx;
             else throw (PyxException) lastEx;
         } finally {
@@ -520,6 +524,7 @@ public class Pyx implements Closeable {
     protected static class PyxResponse {
         private final Response resp;
         private final JSONObject obj;
+        protected List<Exception> exceptions = null;
 
         PyxResponse(@NonNull Response resp, @NonNull JSONObject obj) {
             this.resp = resp;
@@ -549,7 +554,7 @@ public class Pyx implements Closeable {
             this.editable = editable;
         }
 
-        Server(JSONObject obj) throws JSONException {
+        Server(@NonNull JSONObject obj) throws JSONException {
             this(parseUrlOrThrow(obj.getString("uri")), parseNullableUrl(obj.optString("metrics")), obj.getString("name"),
                     obj.has("params") ? new Params(obj.getJSONObject("params")) : Params.defaultValues(),
                     obj.optBoolean("editable", true));
@@ -561,7 +566,7 @@ public class Pyx implements Closeable {
             else return HttpUrl.parse(url);
         }
 
-        static void parseAndSave(@NonNull JSONArray array) throws JSONException {
+        static void parseAndSave(@NonNull JSONArray array, boolean cache) throws JSONException {
             List<Server> servers = new ArrayList<>(array.length());
             for (int i = 0; i < array.length(); i++) {
                 JSONObject obj = array.getJSONObject(i);
@@ -586,7 +591,8 @@ public class Pyx implements Closeable {
                 json.put(server.toJson());
 
             JsonStoring.intoPrefs().putJsonArray(PK.API_SERVERS, json);
-            Prefs.putLong(PK.API_SERVERS_CACHE_AGE, System.currentTimeMillis());
+            if (cache) Prefs.putLong(PK.API_SERVERS_CACHE_AGE, System.currentTimeMillis());
+            else Prefs.putLong(PK.API_SERVERS_CACHE_AGE, 0);
         }
 
         @Nullable