+++ CHANGELOG.md
@@ -4,6 +4,32 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 28-07-2020
+### Added
+- Highlight Overloaded users in player list
+
+### Changed
+- Load default server on network error
+- Better edit game options dialog layout
+
+
+## [4.0.0-beta05] - 26-07-2020
+### Added
+- Added explanation dialog when importing deck
+- Added deck description note
+
+### Fixed
+- Fixed imported decks not syncing properly (Overloaded)
+- Improved sync when requests fail (Overloaded)
+- Correctly close PYX connection
+- Fixed "nickname already in use" error when logging in
+
+
+## [4.0.0-beta04] - 22-07-2020
+### Fixed
+- Fixed chat not working (Overloaded)
+
+
 ## [4.0.0-beta03] - 21-07-2020
 ### Added
 - Added confirmation for game logout